faet: 안건 생성

+++ app/src/main/java/com/hgh/na_o_man/data/dto/agenda/request/AgendaRequestDto.kt
@@ -6,7 +6,7 @@ import com.hgh.na_o_man.domain.model.agenda.AgendaRequestModel
 data class AgendaRequestDto (
     val shareGroupId: Long,
     val title: String,
-    val photoIdList: List<Int>
+    val photoIdList: List<Long>
 ) {
     fun toModel() = AgendaRequestModel(
         shareGroupId = this.shareGroupId,
